python package: Changed directory structure and code so rego wrapper can be used as a python package
