+++ glitch/rego/rego_python/pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "rego_python"
+version = "0.1.0"
+description = "Python wrapper for Rego engine"
+authors = [
+    {name = "Daniel Carvalho", email = "daniel.m.carvalho2002@gmail.com"},
+    {name = "Rodrigo Coelho", email = ""}
+]
+license = {text = "GNU GENERAL PUBLIC LICENSE v3.0"}
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = []
+
+[tool.setuptools.package-data]
+"rego_python" = ["libs/**/*"]
