+++ glitch/rego/rego_python/src/rego_python/__init__.py
@@ -1,0 +1,3 @@
+from .wrapper import run_rego
+
+__all__ = ["run_rego"]
